+++ Assets/Source/Scripts/PotentialField/IWeightComponent.cs
@@ -4,5 +4,5 @@ internal interface IWeightComponent
 {
     internal Transform Transform { get; }
     internal int Weight { get; }
-    internal float Offset { get; }
+    internal int Offset { get; }
 }
